// frontend/src/types/UrlFilters.ts
export enum UrlFilters {
    SEASONS = 'seasons',
    COMPETITIONS = 'comps',
    POSITIONS = 'positions',
    MINUTE_FROM = 'minfrom',
    MINUTE_TO = 'minto',
    MINIMUM_AGE = 'minage',
    MAXIMUM_AGE = 'maxage',
    MINIMUM_HEIGHT = 'minheight',
    MAXIMUM_HEIGHT = 'maxheight',
    PLAYER_NAMES = 'names',
    PLAYER_COUNTRIES = 'c',
    CLUBS_PLAYED_FOR = 'clubspf',
    CLUBS_PLAYED_AGAINST = 'clubspa',
    SUBS_ONLY = 'subonly',
    EARLIEST_SUB_ON_TIME = 'earliestsub',
    LATEST_SUB_ON_TIME = 'latestsub',
    PENALTIES = "penalty",
    HOME_OR_AWAY = "home",
    SORT_BY = "sort",
    SCOPE = "scope",
    MINIMUM_APPEARANCES = "ma",
    MINIMUM_GOALS = "ming",
    MAXIMUM_GOALS = "maxg",
    MINIMUM_ASSISTS = "mina",
    MAXIMUM_ASSISTS = "maxa",
    MINIMUM_GOALS_AND_ASSISTS = "minga",
    MAXIMUM_GOALS_AND_ASSISTS = "maxga",
}